data_analysis_agent/requirements.txt
google-adk
streamlit
pandas
sqlalchemy
psycopg2-binary
google-cloud-bigquery
mysql-connector-python
python-dotenv
altair
plotly
